Revert vendored captun server back to npm package
The scriptThrewException is a Cloudflare runtime behavior, not a captun
bug. No need to vendor.
